+++ src/main/java/solutions/taulien/ribac/server/group/GroupRepository.java
@@ -1,0 +1,42 @@
+package solutions.taulien.ribac.server.group;
+
+import com.google.inject.Inject;
+import io.reactivex.Single;
+import solutions.taulien.ribac.server.DbHelper;
+import solutions.taulien.ribac.server.error.DuplicateCreateError;
+import solutions.taulien.ribac.server.gen.jooq.tables.records.DbGroup;
+import solutions.taulien.ribac.server.util.RepositoryHelper;
+
+import static solutions.taulien.ribac.server.gen.jooq.Tables.GROUP;
+
+public class GroupRepository {
+
+    private final DbHelper dbHelper;
+
+
+
+    @Inject
+    public GroupRepository(DbHelper dbHelper) {
+        this.dbHelper = dbHelper;
+    }
+
+
+
+    public Single<DbGroup> createGroup(String name, String requestId) {
+        return this.dbHelper
+                   .execute(
+                       requestId,
+                       db -> db.insertInto(GROUP)
+                               .set(GROUP.NAME, name)
+                               .returning()
+                               .fetchOne()
+                   )
+                   .onErrorResumeNext(
+                       failure -> Single.error(
+                           RepositoryHelper.mySqlRespondedWithDuplicateEntryError(failure)
+                               ? new DuplicateCreateError("A group already exists with the name '" + name + "'")
+                               : failure
+                       )
+                   );
+    }
+}